test(TheoreticalMaxCalculator): 이론치(상위 난도 패턴을 모두 100% 정확도로 클리어 했을 때 DJPOWER 수치)를 계산해서 반환하는 테스트 추가

+++ src/main/java/com/hokago_memories/domain/song/PatternDifficulty.java
@@ -6,6 +6,6 @@ import jakarta.persistence.Embeddable;
 public record PatternDifficulty(
         int level,
         Double floor,
-        Double rating
+        int rating
 ) {
 }
